+++ scene/assets/images/login/platform.png
