+++ scene/assets/images/logo.png
