+++ scene/src/components/chat-message/index.ts
@@ -1,1 +1,1 @@
-export { default as ChatMessage } from './ChatMessage'
+export { default as ChatMessage } from './temp_ChatMessage'
