+++ scene/src/components/friend-item/index.ts
@@ -1,1 +1,1 @@
-export { default as FriendItem } from './FriendItem'
+export { default as FriendItem } from './temp_FriendItem'
