+++ scene/src/components/friend-item/temp_FriendItem.tsx
