+++ scene/src/components/input-field/index.ts
@@ -1,1 +1,1 @@
-export { default as InputField } from './InputField'
+export { default as InputField } from './temp_InputField'
