+++ scene/src/components/slider/Slider.tsx
